+++ tests/test_address.py
@@ -155,3 +155,28 @@ def test_address_verify_invalid_address(test_client):
         test_client.address.verify(address.id)
 
     assert str(error.value) == "Unable to verify address."
+
+
+@pytest.mark.vcr()
+def test_address_create_verify_carrier(incorrect_address, test_client):
+    """Test creating an address with the `verify_carrier` param."""
+    incorrect_address["verify"] = True
+    incorrect_address["verify_carrier"] = "UPS"
+    address = test_client.address.create(**incorrect_address)
+
+    assert isinstance(address, Address)
+
+    assert address.verifications.delivery.errors[0].message == "Address not found"
+    assert address.verifications.zip4.errors[0].message == "Address not found"
+
+
+@pytest.mark.vcr()
+def test_address_create_and_verify_carrier(incorrect_address, test_client):
+    """Test creating and verifying an address with the `verify_carrier` param."""
+    incorrect_address["verify_carrier"] = "UPS"
+    address = test_client.address.create_and_verify(**incorrect_address)
+
+    assert isinstance(address, Address)
+
+    assert address.verifications.delivery.errors[0].message == "Address not found"
+    assert address.verifications.zip4.errors[0].message == "Address not found"